requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Frontend Dependencies
Flask==3.0.0
requests==2.31.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# AI and Data Processing
google-generativeai==0.3.2
pydantic==2.5.0

# Environment and Configuration
python-dotenv==1.0.0

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: For enhanced text processing
nltk==3.8.1
textstat==0.7.3